polished/finished puzzle rendering
Gave up on corners with a split connection. There is no work around to be found using LineRenderer. I tried a couple vector libraries, but they didn't work or weren't worth the time. So, split corners have a sharp edge.

Also extended strokes to ensure the space of lineWidth/2 extending from a node is always covered, no matter the direction or types of paths.